using sprite alpha value to create a color for sprite drawing

+++ RoguePanda/entity/interface/ISpriteObject.cs
@@ -4,7 +4,7 @@ namespace RoguePanda.entity {
     public interface ISpriteObject {
         string texID { get; set; }
         EntityColor color { get; }
-        int alpha { get; set; }
+        byte alpha { get; set; }
         int width { get; set; }
         int height { get; set; }
         float rotation { get; set; }
